+++ src/main/java/com/payflow/entity/LedgerEntryType.java
@@ -1,0 +1,5 @@
+package com.payflow.entity;
+
+public enum LedgerEntryType {
+	DEBIT, CREDIT
+}